move to env utils, pre-commit
